requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
werkzeug==3.0.1
email-validator==2.1.0.post1
python-jose==3.3.0 

aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
babel==2.16.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Markdown==0.3
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-WTF==1.2.2
frozenlist==1.5.0
gotrue==2.11.1
greenlet==3.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.11.0
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
postgrest==0.19.1
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.2
realtime==2.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
text-unidecode==1.3
typing_extensions==4.12.2
websockets==13.1
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
yarl==1.18.3
